% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_colourise_text.R
\name{ansi_escape}
\alias{ansi_escape}
\title{Create escape ansi code for colored printing}
\usage{
ansi_escape(code)
}
\arguments{
\item{code}{An integer of length 1, the code of the color or an ansi code.}
}
\value{
A character of length 1, the escaped ansi code.
}
\description{
Create escape ansi code for colored printing
}
