% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_colourise_text.R
\name{cat_brown}
\alias{cat_brown}
\title{Print Text in Brown Color}
\usage{
cat_brown(msg, bg = NULL)
}
\arguments{
\item{msg}{The text message to be printed.}

\item{bg}{The background color. Defaults to NULL.}
}
\description{
This function prints text in brown color.
}
