% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_colourise_text.R
\name{cat_colourised}
\alias{cat_colourised}
\title{Colorize and Print Text with Customization}
\usage{
cat_colourised(msg, fg = NULL, bg = NULL)
}
\arguments{
\item{msg}{The text message to be colorized and printed.}

\item{fg}{The foreground color. Defaults to NULL.}

\item{bg}{The background color. Defaults to NULL.}
}
\description{
This function colorizes and prints text with specified foreground and background colors.
}
