% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_colourise_text.R
\name{cat_cyan}
\alias{cat_cyan}
\title{Print Text in Cyan Color}
\usage{
cat_cyan(msg, bg = NULL)
}
\arguments{
\item{msg}{The text message to be printed.}

\item{bg}{The background color. Defaults to NULL.}
}
\description{
This function prints text in cyan color.
}
