% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_colourise_text.R
\name{cat_info}
\alias{cat_info}
\title{Print text in Cyan Color}
\usage{
cat_info(msg, bg = NULL)
}
\arguments{
\item{msg}{The text message to be printed.s.}

\item{bg}{The background color. Defaults to NULL.}
}
\description{
This function prints a message in cyan color.
}
