% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_colourise_text.R
\name{cat_red}
\alias{cat_red}
\title{Print Text in Light Red Color}
\usage{
cat_red(msg, bg = NULL)
}
\arguments{
\item{msg}{The text message to be printed.}

\item{bg}{The background color. Defaults to NULL.}
}
\description{
This function prints text in light red color.
}
