% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_colourise_text.R
\name{cat_warning}
\alias{cat_warning}
\title{Print text in Brown Color}
\usage{
cat_warning(msg, bg = NULL)
}
\arguments{
\item{msg}{The text message to be printed.s.}

\item{bg}{The background color. Defaults to NULL.}
}
\description{
This function prints a message in brown color.
}
