% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_colourise_text.R
\name{cat_yellow}
\alias{cat_yellow}
\title{Print Text in Yellow Color}
\usage{
cat_yellow(msg, bg = NULL)
}
\arguments{
\item{msg}{The text message to be printed.}

\item{bg}{The background color. Defaults to NULL.}
}
\description{
This function prints text in yellow color.
}
