% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{catl}
\alias{catl}
\title{Conditional Print to Console (for tables) based on verbosity level}
\usage{
catl(...)
}
\arguments{
\item{...}{Objects to be printed to the console.}
}
\description{
This function prints the provided content to the console only if the global
verbosity level is equal to or greater than \code{VL_50_TABLES}.
}
