% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{combineLogPaths}
\alias{combineLogPaths}
\title{Return the combined paths of the log directory of the specific sub project and a new path}
\usage{
combineLogPaths(path)
}
\arguments{
\item{path}{the sub directory or file name to add}
}
\value{
A character of length one containing the path to add to the sub project specific log directory.
}
\description{
Return the combined paths of the log directory of the specific sub project and a new path
}
