% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{complete_table}
\alias{complete_table}
\title{Complete Table with Missing Columns}
\usage{
complete_table(table, table_description)
}
\arguments{
\item{table}{A data.table representing the table to be completed.}

\item{table_description}{A description of the expected table structure.}
}
\value{
A completed data.table with missing columns added.
}
\description{
This function completes a table by adding missing columns based on the specified table description.
}
