% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{convertDateFormat}
\alias{convertDateFormat}
\title{Convert Date Format}
\usage{
convertDateFormat(dt, column)
}
\arguments{
\item{dt}{A data table.}

\item{column}{The name of the column whose date format needs to be converted.}
}
\value{
This function modifies the input data table \code{dt} by converting the date format
of the specified column.
}
\description{
This function converts the date format of a column in a data table.
}
\details{
This function takes a data table \code{dt} and the name of a column \code{column}
as input. It converts the date format of the specified column in the following steps:
\itemize{
\item Convert the column to character type.
\item Remove time information (if any) from the column values.
\item Replace '/' with '-' in the column values.
\item If the values match the pattern YYYY, append '-01-01' to represent the complete date.
\item If the values match the pattern YYYY-MM, append '-01' to represent the complete date.
\item Finally, convert the column to date type using \code{lubridate::as_date}.
}
}
