% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_datetime.R
\name{convertDateTimeFormat}
\alias{convertDateTimeFormat}
\title{Fix datetime format in specified columns}
\usage{
convertDateTimeFormat(dt, column)
}
\arguments{
\item{dt}{A data table.}

\item{column}{A character vector specifying the column to fix.}
}
\value{
This function modifies the input data table \code{dt} in place by
fixing the date format in the specified column
}
\description{
This function fixes the date format in specified columns of a data table.
}
\details{
This function expects a data table \code{dt} and a character vector
\code{column} specifying the column to be fixed. It converts the values in
the specified column to date-time objects with the format \code{ymd_hms},
truncating the time to minutes and setting the timezone to "Europe/Berlin".
}
