% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{convertIntegerFormat}
\alias{convertIntegerFormat}
\title{Fix integer format in specified column}
\usage{
convertIntegerFormat(dt, column)
}
\arguments{
\item{dt}{A data table.}

\item{column}{A character vector specifying the column to fix.}
}
\value{
This function modifies the input data table \code{dt} in place by
fixing the integer format in the specified column
}
\description{
This function fixes the integer format in specified column of a data table.
}
\details{
This function expects a data table \code{dt} and a character vector
\code{column} specifying the column to be fixed. It converts the values in
the specified column to integers using \code{as.integer}.
}
