% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_datetime.R
\name{convertTimeFormat}
\alias{convertTimeFormat}
\title{Convert Time Format}
\usage{
convertTimeFormat(dt, column)
}
\arguments{
\item{dt}{A data table.}

\item{column}{The name of the column whose time format needs to be converted.}
}
\value{
This function modifies the input data table \code{dt} by converting the time format
of the specified column.
}
\description{
This function converts the time format of a column in a data table.
}
\details{
This function takes a data table \code{dt} and the name of a column \code{column}
as input. It converts the time format of the specified column using the following steps:
\itemize{
\item Parse the column values as POSIXct objects using \code{strptime}.
\item Format the POSIXct objects to the desired time format using \code{format}.
\item Update the column values in the data table with the formatted time values.
}
}
