% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_clock.R
\name{createClock}
\alias{createClock}
\title{Create a Clock Object}
\usage{
createClock(clock_variable_name = "PROCESS_CLOCK")
}
\arguments{
\item{clock_variable_name}{The name of the variable to which the clock object will be assigned. Default is "PROCESS_CLOCK".}
}
\value{
The clock object is assigned to the global environment under the specified variable name.
}
\description{
This function creates a clock object and assigns it to the global environment under the specified variable name.
}
\details{
The function creates a clock object using the \code{methods::new()} function with the class "Clock" and assigns it to the global environment.
}
