% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{dataTableAsCharacter}
\alias{dataTableAsCharacter}
\title{Convert Data.Table to Character String}
\usage{
dataTableAsCharacter(dt, header = FALSE, footer = FALSE)
}
\arguments{
\item{dt}{A data.table to be converted.}

\item{header}{Logical, indicating whether to include a header with column names.}

\item{footer}{Logical, indicating whether to include a footer with column names.}
}
\value{
A formatted character string representation of the data.table.
}
\description{
This function converts a data.table to a formatted character string with optional header and footer.
}
\details{
The function converts a data.table to a character string, aligning columns and optionally including a header and footer.
It pads each column to the maximum width of its elements for better alignment.
}
