% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_db.R
\name{dbAddContent}
\alias{dbAddContent}
\title{Insert Rows into a PostgreSQL Table}
\usage{
dbAddContent(db_connection, table_name, table)
}
\arguments{
\item{db_connection}{A valid database connection object, typically created using \code{DBI::dbConnect}.}

\item{table_name}{The name of the target table in the PostgreSQL database where rows will be inserted.
The table name is automatically converted to lower case.}

\item{table}{A \code{data.table} or \code{data.frame} containing the rows to be inserted into the specified table.}
}
\description{
This function inserts rows from a \code{data.table} or \code{data.frame} into a specified table within a PostgreSQL database.
It automatically converts the table name to lower case to align with PostgreSQL's case sensitivity conventions.
The function attempts to insert all provided rows into the database and reports the number of rows inserted
along with the duration of the operation. In case of an error during the insert operation, it prints out the error
message and stops execution.
}
