% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_db.R
\name{dbCheckContent}
\alias{dbCheckContent}
\title{Check column widths of a table in a PostgreSQL database}
\usage{
dbCheckContent(db_connection, table_name, table)
}
\arguments{
\item{db_connection}{A DBI connection object to the PostgreSQL database.}

\item{table_name}{A character string specifying the name of the table to check.}

\item{table}{A data frame representing the table to be checked.}
}
\value{
A logical value indicating whether the table content meets the length constraints.
}
\description{
This function retrieves information about the column widths of a specified table
in a PostgreSQL database using the information_schema.columns system view.
}
\details{
This function converts the table name to lowercase for PostgreSQL compatibility,
then constructs a SQL query to retrieve the column widths (maximum length for VARCHAR columns)
of the specified table from the information_schema.columns system view. It retrieves the
column widths using \code{DBI::dbGetQuery()} and checks the lengths of the corresponding
columns in the provided table. If any values exceed the maximum length, it prints an error
message and returns FALSE. Otherwise, it returns TRUE indicating no violations were found.
}
