% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_db.R
\name{dbConnect}
\alias{dbConnect}
\title{Establish a Connection to a PostgreSQL Database}
\usage{
dbConnect(user, password, dbname, host, port, schema)
}
\arguments{
\item{user}{The username for the database connection.}

\item{password}{The password for the database connection.}

\item{dbname}{The name of the database.}

\item{host}{The host address where the database is located.}

\item{port}{The port number for the database connection.}

\item{schema}{The schema under which the tables can be found. This sets the search path to the specified schema.}
}
\value{
A database connection object that is configured to use the specified schema and has an increased \code{work_mem}
setting.
}
\description{
This function establishes a connection to a PostgreSQL database using the specified credentials and settings.
It configures the connection to use a specific schema and adjusts the \code{work_mem} setting to allow more memory for
operations, enhancing performance for tasks that require more memory.
}
