% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_db.R
\name{dbDeleteContent}
\alias{dbDeleteContent}
\title{Delete All Rows from a Database Table}
\usage{
dbDeleteContent(db_connection, table_name)
}
\arguments{
\item{db_connection}{A database connection object created by \code{DBI::dbConnect} or a similar
DBI connection function. This connection should be active and valid for
the deletion to work.}

\item{table_name}{The name of the table from which all rows should be deleted. The table name
will be converted to lower case to comply with PostgreSQL naming conventions.}
}
\value{
An integer value indicating the number of rows affected by the delete operation. For
a successful deletion, this will be the number of rows that were in the table prior
to deletion.
}
\description{
This function deletes all rows from a specified table in the database. It is designed to
work with PostgreSQL databases, where table names are case-sensitive and typically lower case.
The function converts the provided table name to lower case before executing the delete operation.
}
