% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_db.R
\name{dbDisconnect}
\alias{dbDisconnect}
\title{Close Database Connection}
\usage{
dbDisconnect(db_connection)
}
\arguments{
\item{db_connection}{A valid database connection object created by \code{DBI::dbConnect} or
a similar DBI connection function. This connection will be closed by this function.}
}
\value{
Invisible \code{TRUE} if the disconnection was successful, otherwise an error is thrown
by the underlying DBI function.
}
\description{
This function serves as a wrapper around \code{DBI::dbDisconnect}, providing a simplified
interface for closing an open database connection. It ensures that the connection
is properly closed to free up resources.
}
\seealso{
\code{\link[DBI]{dbDisconnect}} for the underlying DBI function.
}
