% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_logging.R
\name{END__}
\alias{END__}
\title{Print Footer of a Block}
\usage{
END__(verbose = VERBOSE - 4, len = 104)
}
\arguments{
\item{verbose}{An integer of length 1, the verbose level. If verb is 0, no output will be produced.}

\item{len}{An integer of length 1, the length of the underline.}
}
\description{
Print Footer of a Block
}
