% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_logging.R
\name{end_logging}
\alias{end_logging}
\title{End Logging}
\usage{
end_logging()
}
\description{
This function finalizes the logging process by closing the log file.
}
\details{
The function stops the redirection of console output and messages to the log file initiated by \code{start_logging}.
It closes the log file, ensuring that no further entries are appended to it.
}
