% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{ETLUtils-package}
\alias{ETLUtils-package}
\alias{ETLUtils}
\title{Extra utility functions to execute standard ETL operations on large data}
\description{
Provides functions to load bigdata (e.g. from Oracle) directly into \code{ffdf} objects using DBI and some utility functions like recoding and matchmerge which does fast left outer join merging based on \code{match}.
}
\examples{
# See the specified functions in the package
}
\author{
Jan Wijffels \email{jwijffels@bnosac.be}
}
