% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{existsLocalRdataFile}
\alias{existsLocalRdataFile}
\title{Check if an RData file exists locally}
\usage{
existsLocalRdataFile(table_name)
}
\arguments{
\item{table_name}{The name of the table.}
}
\value{
TRUE if the RData file exists locally, otherwise FALSE.
}
\description{
This function checks if an RData file corresponding to the specified table exists locally.
}
