% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_medical_codes.R
\name{expandICDs}
\alias{expandICDs}
\title{Expand the given ICD Codes.}
\usage{
expandICDs(
  icdCodes,
  minyear = NA,
  maxyear = NA,
  fullExpanded = TRUE,
  normcodesOnly = TRUE
)
}
\arguments{
\item{icdCodes}{The ICD code to be expanded.}

\item{minyear}{the minimum year to search for the codes in the ICD database. NA is default and means current year.}

\item{maxyear}{the maximum year to search for the codes in the ICD database. NA is default and means current year.}

\item{fullExpanded}{logical. If TRUE then only full expanded codes will be returned.}

\item{normcodesOnly}{logical. Only relevant if fullExpanded is FALSE. If TRUE then codes with bars like 'A21.-' will be returned too.}
}
\value{
A character vector containing the expanded ICD codes.
}
\description{
This function expands ICD codes by retrieving corresponding labels for the provided ICD codes.
It filters the results to include only those codes that have the same prefix as the input ICD codes.
}
\examples{
# Example usage:
print(expandICDs(c('I25', 'A90')))
print(expandICDs('I25', 2018, 2021, fullExpanded = FALSE, normcodesOnly = TRUE))
print(expandICDs('I25', 2018, 2021, fullExpanded = FALSE, normcodesOnly = FALSE))

}
