% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_str.R
\name{extract_words}
\alias{extract_words}
\title{Extract Words from String}
\usage{
extract_words(input_string)
}
\arguments{
\item{input_string}{A character string from which to extract words.}
}
\value{
A character vector containing the extracted words.
}
\description{
Extracts individual words from a string that contains words separated by
any number of spaces and/or commas. Returns a vector of words.
}
\examples{
input_string <- "Here are some words, separated, by, commas, and spaces"
extract_words(input_string)
}
