% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_sample_resources.R
\name{get_ncores}
\alias{get_ncores}
\title{Get the Number of Cores Available for Parallelization}
\usage{
get_ncores(os)
}
\arguments{
\item{os}{A character string representing the operating system name.}
}
\value{
An integer specifying the number of cores available for parallelization.
}
\description{
This function determines the number of CPU cores available for parallelization
based on the operating system.
}
