% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_sample_resources.R
\name{get_os}
\alias{get_os}
\title{Get the Operating System Name}
\usage{
get_os()
}
\value{
A character string representing the operating system name.
}
\description{
This function retrieves the name of the operating system and returns it in lowercase.
}
