% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_misc.R
\name{get_pattern_YYYY}
\alias{get_pattern_YYYY}
\title{Get a regular expression pattern for matching YYYY format}
\usage{
get_pattern_YYYY()
}
\value{
A character vector representing the regular expression pattern for YYYY format.
}
\description{
This function returns a regular expression pattern for matching the YYYY format
(four-digit year). It can be used for validating and extracting year information
from strings.
}
\examples{
get_pattern_YYYY()

}
