% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{get_project_dir_names}
\alias{get_project_dir_names}
\title{Generate directory names for project outputs.}
\usage{
get_project_dir_names(
  project_name = PROJECT_NAME,
  project_time_stamp = PROJECT_TIME_STAMP
)
}
\arguments{
\item{project_name}{The name of the project.}

\item{project_time_stamp}{Time stamp as extension for the current project subfolders}
}
\value{
A named list containing directory names for global and local project outputs.
}
\description{
This function takes a project name and constructs directory names for both global and local project outputs.
It creates directory names for various result categories, including bundles, log, performance, and tables.
The function utilizes predefined global and local output folder names and appends a project-specific timestamp to them.
The resulting directory names are organized into a named list for easy access and reference.
}
