% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_str.R
\name{getAfterLastSlash}
\alias{getAfterLastSlash}
\title{Extract Everything After the Last Slash}
\usage{
getAfterLastSlash(strings)
}
\arguments{
\item{strings}{A character vector where each element is a string potentially
containing slashes.}
}
\value{
A character vector of the same length as \code{strings}, where each element
is the part of the original string after the last slash.
}
\description{
This function takes a vector of strings and returns a new vector where each element
is modified to include only the portion of the original string that appears
after the last slash.
}
\examples{
strings <- c('Patient/PID_001', 'Encounter/EID_001', 'Condition/CID_001', 'aaa/bbb/ccc', 'aaa')
getAfterLastSlash(strings)

}
