% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{getcolumnIndex}
\alias{getcolumnIndex}
\title{Get the index of a column in a data.table by its name.}
\usage{
getcolumnIndex(table, columnName)
}
\arguments{
\item{table}{A data.table.}

\item{columnName}{The name of the column to search for.}
}
\value{
The index of the specified column, or 0 if the column is not found.
}
\description{
This function returns the index of the first column in a data.table that matches
the specified column name.
}
\examples{
library(data.table)

# Create a sample data.table
dt <- data.table(ID = 1:5, Name = c("Alice", "Bob", "Charlie", "David", "Eve"))

# Get the index of the "Name" column
index <- getcolumnIndex(dt, "Name")
print(index)  # Output: [1] 2

# Try to get the index of a non-existing column
index <- getcolumnIndex(dt, "Age")
print(index)  # Output: [1] 0

# Try to get the index of column name NA
index <- getcolumnIndex(dt, NA)
print(index)  # Output: [1] 0

}
