% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{getLocalRdataFileInfo}
\alias{getLocalRdataFileInfo}
\title{Get file information for an RData file}
\usage{
getLocalRdataFileInfo(table_name)
}
\arguments{
\item{table_name}{The name of the table.}
}
\value{
A list containing file information such as size, permissions, and timestamps.
}
\description{
This function retrieves information about an RData file corresponding to the specified table.
}
