% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{getLocalRdataFileName}
\alias{getLocalRdataFileName}
\title{Get the filename for an RData file corresponding to a table}
\usage{
getLocalRdataFileName(table_name)
}
\arguments{
\item{table_name}{The name of the table.}
}
\value{
A character string representing the filename for the RData file.
}
\description{
This function constructs the filename for an RData file corresponding to the specified table.
}
