% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_str.R
\name{greplic}
\alias{greplic}
\title{Perform a case-insensitive pattern search using \code{grepl} with Perl patterns (perl = TRUE).}
\usage{
greplic(pattern, x, whole_word = FALSE, perl = TRUE)
}
\arguments{
\item{pattern}{A pattern to search for.}

\item{x}{A list or table where the pattern should be searched.}

\item{whole_word}{If TRUE, the pattern will be treated as a whole word and matched accordingly.
It adds '^' to the beginning and '$' to the end of the pattern if they don't exist.}

\item{perl}{If TRUE, uses Perl-compatible regular expressions for pattern matching (default is TRUE).}
}
\description{
Perform a case-insensitive pattern search using \code{grepl} with Perl patterns (perl = TRUE).
}
\examples{
pattern <- 'AAA|BBB'
greplic(pattern, 'I am a sentence with AAAA and CCC.', whole_word = TRUE) # FALSE
greplic(pattern, 'I am a sentence with AAA and CCC.', whole_word = TRUE) # TRUE

}
\seealso{
grepl

This function performs case-insensitive pattern searches using \code{grepl} with Perl patterns. If \code{whole_word} is set to TRUE,
the pattern is treated as a whole word, and '^' is added to the beginning and '$' to the end of the pattern if they are missing.
When \code{perl} is TRUE, Perl-compatible regular expressions are used for pattern matching.
}
