% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_clock.R
\name{human_readable_duration}
\alias{human_readable_duration}
\title{Make a duration human readable}
\usage{
human_readable_duration(
  time_in_secs = c(0 * 24 * 3600 + 1 * 3600 + 1 * 60 + 1.001001001, 0 * 24 * 3600 + 1 *
    3600 + 0 * 60 + 1.001001001)
)
}
\arguments{
\item{time_in_secs}{A number of length one.}
}
\value{
A character of length one containing the human readable duration.
}
\description{
Make a duration human readable
}
\examples{
human_readable_duration(time_in_secs = 11 * 24 * 3600 + 1 * 3600 + 1 * 60 + 1.001001001)

}
