% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{initConstants}
\alias{initConstants}
\title{Loads a configuration toml file and sets all variables in this file in the global
context.}
\usage{
initConstants(path_to_toml)
}
\arguments{
\item{path_to_toml}{path to the configuration toml file.}
}
\description{
Loads a configuration toml file and sets all variables in this file in the global
context.
}
