% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_medical_codes.R
\name{interpolar_expandICDs}
\alias{interpolar_expandICDs}
\title{Expand the given ICD Codes between the polar start year and current year.
The codes are expanded with general codes and codes with bars like 'A20.-'}
\usage{
interpolar_expandICDs(...)
}
\arguments{
\item{...}{the ICD codes to be expanded}
}
\description{
Expand the given ICD Codes between the polar start year and current year.
The codes are expanded with general codes and codes with bars like 'A20.-'
}
\seealso{
expandICDs
}
