% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{isDebug}
\alias{isDebug}
\title{Check if debugging mode is enabled.}
\usage{
isDebug()
}
\value{
TRUE if debugging mode is enabled, otherwise FALSE.
}
\description{
This function checks if debugging mode is enabled by verifying the existence of the 'DEBUG' variable and if it is
set to TRUE. his can be used in code to enable code sections only if a global variable DEGUG is set to TRUE.
}
