% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{isOK}
\alias{isOK}
\title{Tests the passed object for being not an error.}
\usage{
isOK(obj)
}
\arguments{
\item{obj}{The object to be tested for not being an error.}
}
\value{
FALSE if the object is an 'try-error', TRUE otherwise.
}
\description{
This function checks whether the input object is not an 'try-error'. If the object
is not an error, the function returns FALSE; otherwise, it returns TRUE.
}
\examples{
# Example 1: Check if an error is not an error
result <- try(log("a"))
isOK(result)

# Example 2: Check if a numeric value is an error
numeric_result <- 42
isOK(numeric_result)

}
