% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{isSimpleFalseOr0}
\alias{isSimpleFalseOr0}
\title{Check if a value is a simple false value (FALSE) or zero (0).}
\usage{
isSimpleFalseOr0(x)
}
\arguments{
\item{x}{The value to be checked.}
}
\value{
TRUE if the input is a simple false value or zero, FALSE otherwise.
}
\description{
This function evaluates whether the input value is an atomic element, has a length of 1,
and is either a logical false value (FALSE) or zero (0). If these conditions are met,
the function returns TRUE; otherwise, it returns FALSE.
}
\examples{
# Example 1: Check if FALSE is a simple false value (TRUE)
isSimpleFalseOr0(FALSE)

# Example 2: Check if 0 is a simple false value  (TRUE)
isSimpleFalseOr0(0)

# Example 3: Check if 1 is not a simple false value (FALSE)
isSimpleFalseOr0(1)

# Example 4: Check if a boolean vector is a simple true value (FALSE)
isSimpleTrueOrNot0(c(FALSE, FALSE))

}
