% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_sample_resources.R
\name{limit_ncores}
\alias{limit_ncores}
\title{Limit the Number of Cores Based on Available Cores}
\usage{
limit_ncores(ncores)
}
\arguments{
\item{ncores}{An integer specifying the desired number of cores. If NULL, defaults to 1.}
}
\value{
An integer representing the limited number of cores.
}
\description{
This function limits the number of cores based on the available cores and a specified limit.
}
