% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_sample_resources.R
\name{loadFHIRResourcesByOwnID}
\alias{loadFHIRResourcesByOwnID}
\title{Load Resources by Their Own IDs}
\usage{
loadFHIRResourcesByOwnID(ids, table_description)
}
\arguments{
\item{ids}{A vector of resource IDs. These should be the unique identifiers for the resources you
wish to download and process. The function expects these IDs to be in a specific format, where
the actual ID follows the last slash ("/") in each string.}

\item{table_description}{An object containing the description of the table where the resources
are to be found. This object must have a specific structure, including a resource slot that
contains the URL or identifier needed for downloading the resources.}
}
\value{
Returns a table of the downloaded resources, processed and cracked open according to the
specifications in \code{table_description}. The exact structure of the returned table depends on the
\code{table_description} parameter and the data processing within \code{downloadAndCrackFHIRResourcesByPIDs}.
}
\description{
This function downloads and processes resources specified by their IDs. It uses a parallel
downloading and cracking method tailored for handling resources described in a table structure.
The function is part of a larger workflow designed to work with polar data resources.
}
