% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_sample_resources.R
\name{loadFHIRResourcesByPID}
\alias{loadFHIRResourcesByPID}
\title{Load Resources by Patient IDs}
\usage{
loadFHIRResourcesByPID(patient_IDs, table_description)
}
\arguments{
\item{patient_IDs}{A vector of patient IDs, which are the unique identifiers for the patients
whose resources you wish to download and process.}

\item{table_description}{An object describing the table where the resources are to be found.
This object must have a specific structure, including a resource slot that contains the URL or
identifier needed for downloading the resources.}
}
\value{
Returns a table of the downloaded resources, processed according to the specifications
in \code{table_description}. The function handles different types of resources by adapting the ID
parameter string based on the resource type, ensuring correct processing.
}
\description{
This function is designed to load resources based on patient IDs. It checks if the requested
resource is "Patient" and uses a specific loading function or the general downloading and
cracking function for other types of resources. It's tailored for handling resources described
in a table structure, specifically for patient-related data.
}
