% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_sample_resources.R
\name{loadResourcesByPID}
\alias{loadResourcesByPID}
\title{Download FHIR resources by patient IDs and perform parallel cracking for each resource type.}
\usage{
loadResourcesByPID(patientIDs, table_description)
}
\arguments{
\item{patientIDs}{A vector of patient IDs for whom FHIR resources should be retrieved.}

\item{table_description}{A list of table descriptions for different FHIR resource types.}
}
\value{
A list containing tables for each resource type, with resource type names as keys.
}
\description{
This function iterates over the resource types defined in table_description, and for each resource type,
it calls the polar_download_by_ids_and_crack_parallel function to download and crack FHIR resources
associated with the given patient IDs. The download behavior is adjusted based on the resource type.
}
