% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_misc.R
\name{log_run}
\alias{log_run}
\title{Runs the given function surrounded with \code{\link[=START__]{START__()}} and \code{\link[=END__]{END__()}}.}
\usage{
log_run(process, verbose = VERBOSE - 4, len = 104)
}
\arguments{
\item{process}{a function}

\item{verbose}{An integer of length 1, the verbose level. If verb is 0, no output will be produced.}

\item{len}{An integer of length 1, the length of the underline.}
}
\description{
Runs the given function surrounded with \code{\link[=START__]{START__()}} and \code{\link[=END__]{END__()}}.
}
