% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{mergeMultipleSpaces}
\alias{mergeMultipleSpaces}
\title{Merge Multiple Whitespace Characters into a Single Space}
\usage{
mergeMultipleSpaces(string)
}
\arguments{
\item{string}{The input string to process.}
}
\value{
The modified string with multiple whitespace characters merged into
a single space.
}
\description{
This function replaces sequences of one or more whitespace characters
(spaces, tabs, newlines, etc.) in the provided string with a single space,
effectively merging multiple spaces into one.
}
\examples{
mergeMultipleSpaces("This   is  a   test.")
}
