% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{moveColumnBefore}
\alias{moveColumnBefore}
\title{Move a column in a data.table directly before another column based on indices}
\usage{
moveColumnBefore(dt, column_to_move, target_column)
}
\arguments{
\item{dt}{A data.table object.}

\item{column_to_move}{The name of the column to move.}

\item{target_column}{The name of the column before which to place the moved column.}
}
\value{
The data.table with adjusted column order. The original data.table is modified by reference.
}
\description{
This function takes a data.table, the name of the column to be moved, and
the name of the target column before which the first column should be placed.
It adjusts the order of the columns in the data.table accordingly by manipulating
column indices, without altering other columns' order unnecessarily.
}
\examples{
library(data.table)
dt <- data.table(x = 1:3, y = 4:6, z = 7:9)
moveColumnBefore(dt, "y", "z")
print(dt) # Should show columns in the order x, z, y

}
