% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_collections.R
\name{namedListByValue}
\alias{namedListByValue}
\title{Create a named list based on the values of a vector.}
\usage{
namedListByValue(...)
}
\arguments{
\item{...}{A vector from which a named list will be created.}
}
\value{
A named list where the names are identical to the values of the input vector.
}
\description{
This function takes a vector as input and returns a named list in which the names and the content of the list elements are identical.
}
\examples{
# Create a named list from a vector
values <- c("apple", "banana", "cherry", NA_character_)
named_list <- namedListByValue(values)
print(named_list)

}
