% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_misc.R
\name{plural_s}
\alias{plural_s}
\title{Pluralize Suffix Based on Count}
\usage{
plural_s(counts)
}
\arguments{
\item{counts}{Numeric vector of counts.}
}
\value{
An empty string for count 1, 's' otherwise.
}
\description{
This function returns an empty string for count 1 and 's' for any other count.
}
