% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_sample_resources.R
\name{polar_download_by_ids_and_crack_parallel}
\alias{polar_download_by_ids_and_crack_parallel}
\title{Downloads and cracks FHIR resources in parallel for a given resource type and patient IDs.}
\usage{
polar_download_by_ids_and_crack_parallel(
  resource,
  ids,
  table_description,
  ids_at_once = IDS_AT_ONCE,
  id_param_str,
  verbose = VERBOSE
)
}
\arguments{
\item{resource}{The type of FHIR resource to download and crack.}

\item{ids}{A vector of patient IDs for the FHIR resources.}

\item{table_description}{A table description object specifying the structure of the resulting data table.}

\item{ids_at_once}{The maximum number of IDs to process in each iteration (default: IDS_AT_ONCE).}

\item{id_param_str}{Additional parameter string for constructing the FHIR request URL (default VERBOSE).}

\item{verbose}{Verbosity level}
}
\value{
A data.table containing the cracked FHIR resources.
}
\description{
This function downloads FHIR resources specified by the provided IDs for a given resource type in parallel.
It then performs cracking on the obtained bundles using the specified table description.
}
