% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_sample_resources.R
\name{polar_get_resources_ids}
\alias{polar_get_resources_ids}
\title{Get Resources' IDs}
\usage{
polar_get_resources_ids(endpoint, resource, parameters = NULL, verbose = 0)
}
\arguments{
\item{endpoint}{A character of length 1 containing the FHIR R4 endpoint.}

\item{resource}{A character of length 1 containing the Resource's name.}

\item{parameters}{Either a character of length 1 containing the parameter string or
a named vector of the form c("gender" = "male", "_summary" = "count").}

\item{verbose}{An integer of length 1 containing the level of verbosity. Defaults to 0.}
}
\value{
A character containing the IDs all requested Resources.
}
\description{
Get Resources' IDs
}
