% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{polar_load_bundle}
\alias{polar_load_bundle}
\title{Load a single Bundle}
\usage{
polar_load_bundle(path)
}
\arguments{
\item{path}{the path to load from}
}
\value{
A fhir_bundle as XML
}
\description{
polar_load('bundle.xml')
}
