% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{polar_path_to_bundles_directory}
\alias{polar_path_to_bundles_directory}
\title{Return the path to the bundles directory of the specific sub project}
\usage{
polar_path_to_bundles_directory()
}
\value{
A character of length one containing the path to the sub project specific log directory.
}
\description{
Return the path to the bundles directory of the specific sub project
}
