% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{polar_read_rdata}
\alias{polar_read_rdata}
\title{Read an Object as RDS-File from the \emph{private} \code{tables} directory to which was created for the specific subproject.}
\usage{
polar_read_rdata(filename_without_extension, project_sub_dir = NA)
}
\arguments{
\item{filename_without_extension}{If the default NA is not changed then the name is the name of the table variable.}

\item{project_sub_dir}{subdirectory of the current working directory where the table is located. If NA (default),
then the tables will be loaded from outputLocal/tables'.}
}
\value{
the object
}
\description{
a <- polar_read_rdata('a')
}
