% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_sample_resources.R
\name{polar_sample_identified_resources}
\alias{polar_sample_identified_resources}
\title{Sample Identified FHIR Resources}
\usage{
polar_sample_identified_resources(
  endpoint,
  resource,
  ids,
  parameters = NULL,
  sample_size = 20,
  seed = as.double(Sys.time()),
  verbose = 1
)
}
\arguments{
\item{endpoint}{The FHIR server endpoint URL.}

\item{resource}{The name of the FHIR resource to sample.}

\item{ids}{A vector of identifiers for the FHIR resources.}

\item{parameters}{Additional parameters for the FHIR resource query (optional).}

\item{sample_size}{The size of the sample to retrieve.}

\item{seed}{The seed for the random number generator to ensure reproducibility (optional).}

\item{verbose}{The level of verbosity for logging (default is 1).}
}
\value{
A list of sampled FHIR resources.
}
\description{
This function samples identified FHIR resources by randomly selecting a subset of IDs.
It retrieves the sampled resources using the provided endpoint, resource, and parameters.
}
