% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{polar_save_error}
\alias{polar_save_error}
\title{Save an Error Message in the \emph{private} \code{log} directory to which was created for the specific subproject.}
\usage{
polar_save_error(err, filename_without_extension)
}
\arguments{
\item{err}{A character of length one. The error message.}

\item{filename_without_extension}{A character of length one.}
}
\value{
Nothing.
}
\description{
polar_save_error(err = "Housten! We have a Problem!", filename_without_extension = "big_problem")
}
