% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{polar_save_request}
\alias{polar_save_request}
\title{Save a Request without url and endpoint in the \emph{public} \code{requests} directory to which was created for the specific subproject.}
\usage{
polar_save_request(request, filename_without_extension)
}
\arguments{
\item{request}{A character of length one holding the fhir search request with url and endpoint.}

\item{filename_without_extension}{A character of length one.}
}
\value{
Nothing.
}
\description{
(request <- fhircrackr::fhir_url(
url        = "http://vonk.fire.ly/R4",
resource   = "Patient",
parameters = c(
"gender" = "female",
"_count" = "10"
)))
polar_save_request(request = request, filename_without_extension = "my_request")
}
