% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{polar_save_result_as_rdata}
\alias{polar_save_result_as_rdata}
\title{Save a Result of an \emph{\strong{ANALYSIS-SCTIPT}} in the \emph{public} \code{results} directory to which was created for the specific subproject.}
\usage{
polar_save_result_as_rdata(..., filename_without_extension)
}
\arguments{
\item{...}{A list of data.}

\item{filename_without_extension}{If the default NA is not changed then the name is the name of the table variable.}
}
\value{
Nothing.
}
\description{
Save a Result of an \emph{\strong{ANALYSIS-SCTIPT}} in the \emph{public} \code{results} directory to which was created for the specific subproject.
}
