% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{polar_save_table_as_csv}
\alias{polar_save_table_as_csv}
\title{Save a Table/Data.Frame as .csv in the \emph{private} \code{tables} directory to which was created for the specific subproject.}
\usage{
polar_save_table_as_csv(table, filename_without_extension = NA, sep = ";")
}
\arguments{
\item{table}{A data.frame/data.table object.}

\item{filename_without_extension}{If the default NA is not changed then the name is the name of the table variable.}

\item{sep}{the field separator string. Values within each row of x are separated by this string.}
}
\value{
Nothing.
}
\description{
d <- data.frame(x=1:10, y=rnorm(10))
polar_save_table_as_csv(table = d)
}
