% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{print_table}
\alias{print_table}
\title{Print a table if VERBOSE level allows.}
\usage{
print_table(table, table_name = NA)
}
\arguments{
\item{table}{The input table to print. For example, you can use the mtcars dataset.}

\item{table_name}{A table name to display in the output. If NA the variable name will be displayed.}
}
\description{
This function prints a summary for the specified table if the VERBOSE level is
greater than or equal to VL_50_TABLES (= 5). It uses the \code{print_table_summary} function
for generating the summary.
}
\details{
This function checks the VERBOSE level (assumed to be a global variable) and
prints a summary for the specified table only if the VERBOSE level is greater than
or equal to VL_50_TABLES. The table_name is obtained from the calling function's name.
}
\examples{
# Load required packages
library(datasets)
library(data.table)

#' # Load the mtcars dataset and convert it to a data.table
data(mtcars)
setDT(mtcars)

# Set VL_50_TABLES and VERBOSE to appropriate values
VL_50_TABLES <- 5
VERBOSE <- 7

# Assuming VERBOSE and VL_50_TABLES are defined
print_table(mtcars)

# Assuming VERBOSE and VL_60_ALL_TABLES are defined
print_table(mtcars, 'This is a table full of cars')

}
\seealso{
\code{\link{print_table_summary}}
}
