% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{print_table_summary}
\alias{print_table_summary}
\title{Print a summary for a table}
\usage{
print_table_summary(table, table_name = "")
}
\arguments{
\item{table}{The input table to summarize. For example, you can use the mtcars dataset.}

\item{table_name}{An optional name for the table, used in the summary output.}
}
\value{
This function does not explicitly return a value. It prints the summary to the console.
}
\description{
This function prints a summary for the specified table, including information
about the class, type, number of available values, and number of missing values
for each column. The summary is displayed in a formatted table.
}
\examples{
# Load required packages
library(datasets)
library(data.table)

# Load the mtcars dataset and convert it to a data.table
data(mtcars)
setDT(mtcars)

# Print summary for the mtcars table
print_table_summary(table = mtcars, table_name = 'mtcars')

}
