% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_clock.R
\name{printClock}
\alias{printClock}
\title{Print Clock Variable}
\usage{
printClock(clock_variable_name = "PROCESS_CLOCK")
}
\arguments{
\item{clock_variable_name}{The name of the clock variable to print. Defaults to "PROCESS_CLOCK".}
}
\description{
This function prints the value of a clock variable based on the provided name.
}
\seealso{
get
}
