% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{replaceNonBreakingSpacesByWhitespace}
\alias{replaceNonBreakingSpacesByWhitespace}
\title{Replace Non-Breaking Spaces with Normal Spaces and Optionally Merge Multiple Spaces}
\usage{
replaceNonBreakingSpacesByWhitespace(string)
}
\arguments{
\item{string}{The input string to process.}
}
\value{
The processed string with non-breaking spaces replaced by normal spaces, and optionally,
multiple whitespace characters merged into a single space.
}
\description{
This function replaces non-breaking spaces (UTF-8 code A0) in the input string with normal spaces.
}
\examples{
string <- "This is a test\u00A0\u00A0string"
print(string)
utf8ToInt(string)
string <- replaceNonBreakingSpacesByWhitespace(string)
utf8ToInt(string)
}
