% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{replacePatternsInString}
\alias{replacePatternsInString}
\title{Replace Patterns in a String with Specified Replacements}
\usage{
replacePatternsInString(
  patternsAndReplacements,
  string,
  ignore.case = FALSE,
  perl = FALSE
)
}
\arguments{
\item{patternsAndReplacements}{A named list where names are patterns to be replaced and values are
the corresponding replacements.}

\item{string}{The string within which the patterns should be replaced.}

\item{ignore.case}{logical. Indicates whether the strings should be compared case sensitive. If TRUE
then all result strings are in lower case.}

\item{perl}{logical. Should Perl-compatible regexps be used?}
}
\value{
The modified string with all specified patterns replaced by their corresponding replacements.
}
\description{
This function replaces specified patterns in a given string with their corresponding replacements.
It takes a named list (\code{patternsAndReplacements}) where each name-value pair corresponds to a pattern
and its replacement. The function performs a case-insensitive replacement of these patterns within
the provided string.
}
\examples{
patternsAndReplacements <- list("hello" = "hi", "world" = "earth")
replacePatternsInString(patternsAndReplacements, "Hello World!")
# Returns "hi earth!"
}
