% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{returnPathToBundlesDir}
\alias{returnPathToBundlesDir}
\title{Return the path to the bundles directory of the specific sub project}
\usage{
returnPathToBundlesDir()
}
\value{
A character of length one containing the path to the sub project specific log directory.
}
\description{
Return the path to the bundles directory of the specific sub project
}
