% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{returnPathToLogDir}
\alias{returnPathToLogDir}
\title{Return the path to the log directory of the specific sub project}
\usage{
returnPathToLogDir(project_name = PROJECT_NAME)
}
\arguments{
\item{project_name}{name of the project}
}
\value{
A character of length one containing the path to the sub project specific log directory.
}
\description{
Return the path to the log directory of the specific sub project
}
