% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_logging.R
\name{run_in_in_ignore_error}
\alias{run_in_in_ignore_error}
\title{Execute an inner script info with a specified message and process}
\usage{
run_in_in_ignore_error(message, process)
}
\arguments{
\item{message}{A character string describing the purpose of the inner script info.}

\item{process}{A function representing the inner script info to be executed.}
}
\description{
This function runs an inner script info with the provided message and process, controlling
the verbosity level. If an error occurs, it is ignored.
}
