% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_logging.R
\name{run_out}
\alias{run_out}
\title{Execute an outer script with a specified message and process}
\usage{
run_out(message, process)
}
\arguments{
\item{message}{A character string describing the purpose of the outer script.}

\item{process}{A function representing the outer script to be executed.}
}
\description{
This function runs an outer script with the provided message and process, controlling
the verbosity level.
}
