% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{runFunctionOnTableValues}
\alias{runFunctionOnTableValues}
\title{Apply a Function to Table Values by Columns}
\usage{
runFunctionOnTableValues(dt, cell_fun, colnames = NA)
}
\arguments{
\item{dt}{A data.table or data.frame to process.}

\item{cell_fun}{The function to apply to each cell in the specified columns. This function should accept a
single character value and return a modified value.}

\item{colnames}{A vector of column names to apply the function to. If NA (default), the function is applied
to all columns.}
}
\value{
The modified data.table or data.frame with the function applied to the specified columns.
}
\description{
This function applies a specified function to each cell in one or more columns of a data.table or data.frame.
It is designed to work with character columns, applying the function only to those columns. The table is
modified in place if it is a data.table. If the input is a data.frame, it is temporarily converted to
data.table for processing, then converted back before returning.
}
\examples{
library(data.table)
dt <- data.table(name = c(" John ", "Jane", "Doe "), age = c(30, 25, 22))
runFunctionOnTableValues(dt, colnames = "name", cell_fun = trimws)
}
