% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_logging.R
\name{runs_in_in}
\alias{runs_in_in}
\title{Execute an inner script info with specified message and process (single line)}
\usage{
runs_in_in(message, process)
}
\arguments{
\item{message}{A character string describing the purpose of the inner script info.}

\item{process}{A function representing the inner script info to be executed.}
}
\description{
This function runs an inner script info with the provided message and process, controlling
the verbosity level. Unlike \code{runs}, this function always displays output in a
single line, regardless of the global verbosity setting.
}
