% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{save_performance}
\alias{save_performance}
\title{Save a Clock history in the \emph{public} \code{performance} directory to which was created for the specific subproject.}
\usage{
save_performance(
  filename_without_extension,
  clock = if (is.null(PROCESS_CLOCK)) NULL else PROCESS_CLOCK
)
}
\arguments{
\item{filename_without_extension}{A character vector of length one.}

\item{clock}{A \code{Clock} object. Defaults to the global environment \code{Clock} object \code{clock_}.}
}
\value{
Nothing.
}
\description{
clock_$reset()
clock_$measure_process_time("Some Process", {for(i in 1:1000)function(n)mean(cos(1:n))})
clock_
polar_save_performance(filename_without_extension = "some_process")
my_clock <- clock()
my_clock$measure_process_time("Some different Process", {for(i in 1:1000)function(n)mean(cos(1:n))})
my_clock
save_performance(filename_without_extension = "some_differen_process", clock = my_clock)
}
