% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{sortListByName}
\alias{sortListByName}
\title{Sort a named list alphabetically by its names.}
\usage{
sortListByName(list)
}
\arguments{
\item{list}{A named list to be sorted.}
}
\value{
A sorted named list.
}
\description{
This function takes a named list and sorts it alphabetically based on its names.
}
\examples{
# Create an unsorted named list
unsorted_list <- list(b = 3, a = 2, c = 1)

# Sort the named list
sorted_list <- sortListByName(unsorted_list)
print(sorted_list)

}
