% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{sortListByValue}
\alias{sortListByValue}
\title{Sort a named list alphabetically by its values.}
\usage{
sortListByValue(list)
}
\arguments{
\item{list}{A list to be sorted.}
}
\value{
A sorted list.
}
\description{
This function takes a list and sorts it alphabetically based on its values.
}
\examples{
# Create an unsorted named list
unsorted_list <- list(b = 3, a = 2, c = 1)

# Sort the named list
sorted_list <- sortListByValue(unsorted_list)
print(sorted_list)

}
