% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_misc.R
\name{START__}
\alias{START__}
\title{START__}
\usage{
START__(verbose = VERBOSE - 4, len = 104)
}
\arguments{
\item{verbose}{An integer of length 1, the verbose level. If verb is 0, no output will be produced.}

\item{len}{An integer of length 1, the length of the underline.}
}
\description{
Print Header of Block
}
