% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_logging.R
\name{start_logging}
\alias{start_logging}
\title{Start Logging to a File}
\usage{
start_logging(prefix)
}
\arguments{
\item{prefix}{A prefix to be used in the log file name.}
}
\description{
This function initializes logging by creating or opening a log file for writing.
}
\details{
The function creates or opens a log file with a name generated from the provided prefix and the current timestamp.
It redirects the console output and messages to the log file, allowing for systematic logging of activities.
The log file is stored in the log directory, and each log entry is appended to the existing file.
}
