% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_logging.R
\name{startProcess}
\alias{startProcess}
\title{Start a Process with Error Handling}
\usage{
startProcess(process)
}
\arguments{
\item{process}{The function representing the process to be executed.}
}
\value{
None (prints clock information and handles errors)
}
\description{
This function initiates a process using the specified function and includes error handling.
}
\seealso{
START__, printClock, END__, stopOnError
}
