% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{stopOnError}
\alias{stopOnError}
\title{Stop on Error}
\usage{
stopOnError(...)
}
\arguments{
\item{...}{Character vectors to be concatenated and printed as an error message.}
}
\description{
This function stops execution and prints the concatenated error message.
}
