% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{trimTableValues}
\alias{trimTableValues}
\title{Trim whitespace from values in a data.table or data.frame column}
\usage{
trimTableValues(dt, colnames = NA)
}
\arguments{
\item{dt}{A data.table or data.frame containing the target column(s).}

\item{colnames}{(Optional) A character vector of column names to trim.
If not specified, all columns in the data.table will be trimmed.}
}
\value{
The updated data.table or data.frame with leading and trailing whitespace trimmed
from the specified column(s). In case of data.table the table will be changed by
reference.
}
\description{
This function trims leading and trailing whitespace from all values
in a specified column or multiple columns of a data.table.
}
\examples{
library(data.table)

df <- data.frame(
  col1 = c(1:4, NA),
  col2 = c(' A', 'B ', ' C ', ' ', NA)
)
print(dt)
df <- trimTableValues(df) # data.frame with assignment (pass by copy)
print(dt)

# Example data.table
dt <- data.table(
  col1 = c(1:4),
  col2 = c(' A', 'B ', ' C ', ' ')
)
print(dt)
trimTableValues(dt)  # data.table without assignment (pass by reference)
print(dt)

dt <- data.table(
  Column1 = c("  Hello  ", "  World  "),
  Column2 = c("  Test  ", "  Example  ")
)

# Call the function to trim values in the "Column1" column
print(dt)
trimTableValues(dt, colnames = "Column1")
print(dt)

}
