% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{writeExcelFile}
\alias{writeExcelFile}
\title{Write data to an Excel file using \code{openxlsx} package}
\usage{
writeExcelFile(tables, file_name)
}
\arguments{
\item{tables}{A \code{data.frame} or a (named) list of objects that can be handled by \code{writeData()}
or \code{writeDataTable()} from the \code{openxlsx} package to write to an Excel file.}

\item{file_name}{A string specifying the file path where the xlsx file will be saved.}
}
\value{
None
}
\description{
This function allows writing a data.frame or a list of objects to an Excel file.
The objects in the list should be compatible with \code{writeData()} or \code{writeDataTable()}
functions from the \code{openxlsx} package. It uses \code{write.xlsx} from \code{openxlsx} to perform the write operation.
}
\seealso{
\code{\link[openxlsx]{write.xlsx}} for the underlying function used to write Excel files.
}
