% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_clock.R
\name{createClock}
\alias{createClock}
\title{Create a Clock}
\usage{
createClock()
}
\value{
a Clock
}
\description{
Create a Clock
}
\examples{
my_clock <- createClock()
my_clock$measure_process_time('Process A', {Sys.sleep(1)})
# show raw data
my_clock$complete()
# reset clock
my_clock$reset()
# show clock
my_clock$show() # the same as my_clock or print(my_clock)
### save time measurements as tsv
# my_clock$write('my_clock')
### save time measurements as csv
# my_clock$write(
#   filename_without_extension = 'my_clock',
#   sep = ',',
#   ext = 'csv',
#   hide_errors = FALSE
# )

}
