\name{ETLUtils-package}
\title{Extra utility functions to execute standard ETL operations on large data}
\description{Extra utility functions to execute standard ETL operations on large data}
\details{\tabular{ll}{
Package: \tab ETLUtils\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-03-26\cr
License: \tab GPL-2\cr
Depends: \tab zoo, DBI, ff\cr
Suggests: \tab RSQLite\cr
LazyLoad: \tab yes\cr
}

Provides functions to do fast left outer join merging based on match, loading of bigdata (e.g. from Oracle) directly into ffdf objects using DBI and some utility functions like recoding.}
\alias{ETLUtils}
\docType{package}
\author{Jan Wijffels \email{jwijffels@bnosac.be}}
\keyword{package}
\examples{# See the specified functions in the package}
