\docType{package}
\name{ETLUtils-package}
\alias{ETLUtils}
\alias{ETLUtils-package}
\title{Extra utility functions to execute standard ETL operations on large data}
\description{
  Provides functions to load bigdata (e.g. from Oracle)
  directly into \code{ffdf} objects using DBI and some
  utility functions like recoding and matchmerge which does
  fast left outer join merging based on \code{match}.
}
\examples{
# See the specified functions in the package
}
\author{
  Jan Wijffels \email{jwijffels@bnosac.be}
}

