% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etlutils-package.R
\docType{package}
\name{etlutils-package}
\alias{etlutils}
\alias{etlutils-package}
\title{etlutils: These are the specific utils for ETL processing}
\description{
This package was created for shared functions as utils in ETL processing contexts (file, database, table, etc).
}
\author{
\strong{Maintainer}: Alexander Struebing \email{alexander.struebing@imise.uni-leipzig.de} (\href{https://orcid.org/0000-0001-5151-7665}{ORCID})

Authors:
\itemize{
  \item Florian Schmidt \email{florian.schmidt@imise.uni-leipzig.de} (\href{https://orcid.org/0000-0003-2027-8213}{ORCID})
}

}
\keyword{internal}
