% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_envir.R
\name{initConstants}
\alias{initConstants}
\title{Loads a configuration toml file and sets all variables in this file in the global
context.}
\usage{
initConstants(path_to_toml, defaults = c())
}
\arguments{
\item{path_to_toml}{path to the configuration toml file.}

\item{defaults}{a vector with default values. The names of the elements are the variables names. If the variable
is not found in the global context after the initialization with the toml file, then this default
values will be set.}
}
\description{
Loads a configuration toml file and sets all variables in this file in the global
context.
}
