% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{naLOCFPlusone}
\alias{naLOCFPlusone}
\title{Performs NA replacement by last observation carried forward but adds 1 to the last observation carried forward.}
\usage{
naLOCFPlusone(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector where NA's are replaced with the LOCF + 1
}
\description{
Performs NA replacement by last observation carried forward but adds 1 to the last observation carried forward. \cr
}
\examples{
require(zoo)
x <- c(2,NA,NA,4,5,2,NA)
naLOCFPlusone(x)
}
\seealso{
\code{\link[zoo]{na.locf}}
}

