% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{stopOnError}
\alias{stopOnError}
\title{Stop on Error}
\usage{
stopOnError(potential_error)
}
\arguments{
\item{potential_error}{The object to check for being an error.}
}
\description{
This function checks if the provided argument is an error. If it is, the
function stops the execution.
}
\seealso{
stop
}
